Rename NexmoSmsClient to SmsClient
