+++ src/examples/java/com/nexmo/client/examples/FlierCode.java
@@ -24,9 +24,9 @@ package com.nexmo.client.examples;
 import com.nexmo.client.NexmoClient;
 import com.nexmo.client.auth.JWTAuthMethod;
 import com.nexmo.client.auth.TokenAuthMethod;
+import com.nexmo.client.sms.SmsClient;
 import com.nexmo.client.voice.Call;
-import com.nexmo.client.messaging.NexmoSmsClient;
-import com.nexmo.client.messaging.messages.TextMessage;
+import com.nexmo.client.sms.messages.TextMessage;
 import com.nexmo.client.verify.VerifyClient;
 import com.nexmo.client.verify.VerifyResult;
 
@@ -46,7 +46,7 @@ public class FlierCode {
         String toNumber = "447700900275";
 
         // Sending an SMS
-        NexmoSmsClient smsClient = new NexmoClient(new TokenAuthMethod(apiKey, apiSecret)).getSmsClient();
+        SmsClient smsClient = new NexmoClient(new TokenAuthMethod(apiKey, apiSecret)).getSmsClient();
         smsClient.submitMessage(new TextMessage(
                 fromNumber, toNumber, "Hello from Nexmo!"
         ));
