+++ src/main/java/com/nexmo/client/sms/messages/parameters/MessageClass.java
@@ -19,7 +19,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.nexmo.client.messaging.messages.parameters;
+package com.nexmo.client.sms.messages.parameters;
 
 
 /**
