+++ src/test/java/com/nexmo/client/sms/callback/AbstractMOServletTest.java
@@ -19,9 +19,9 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.nexmo.client.messaging.callback;
+package com.nexmo.client.sms.callback;
 
-import com.nexmo.client.messaging.callback.messages.MO;
+import com.nexmo.client.sms.callback.messages.MO;
 import org.junit.Test;
 
 import javax.servlet.ServletException;
